Split-out bambu-specific elements to their own 3MF 'variant'.
part of CURA-12099
